Inline jira:: macro
Instead of having to write something like the following in release
notes:

 * https://issues.couchbase.com/browse/PYCBC-849[PYCBC-849]:

Now we can use a macro:

 * jira::PYCBC-849[]

Full usage details and configuration in the supplied `extensions.adoc`
(not added to nav for now, as it's internal. Is there a better place
to put this?)

NB: the configuration is currently done inside the asciidoc extension,
as these don't accept config.
A better solution might be to convert this to an *Antora* extension
(which does accept config) which simply reads the config, and registers
a properly configured asciidoc extension dynamically?