wrap-asciidoc-extension (EXPERIMENTAL)
This is an experimental solution to the problem of wrapping an asciidoc
config from Antora.

This pushes a modified `register` function onto the asciidoc config,
with the requested configuration curried into its lexical scope.

(asciidoc-loader will later call the register function on this config).

I still need confirmation from @mojavelinux as:

 * not sure if this is the expected API, or I'm using implementation
   details?

 * oddity remains around requiring the asciidoc extension in that we
   have to `..` the path to find it.
   (I'd have expected the PATHS to be the same, as both antora and
   asciidoc config setups have the root directory in there, but for
   some reason while inside the antora extension, only `./node_modules`
   is on the path.