replace yml with json
Though openapi spec defaults to yaml, yaml's pathological
flexibility with references means that, even with
`redocly bundle --defererenced` the definition of schemas
is spread across the file.

Outputting as JSON means that you get a flatter (and much
larger file). The advantage is that it's easier to fillet
(e.g. remove all the bits that aren't in the database schema)
to make it a smaller example file for docs-site.