Template block macro (handlebars) (#671)
* Template block macro (handlebars)

This extension enables a `[template]` block which formats a
specified .json file against a .hbs Handlebars template.

```
[template,example$some-data.json]
--
include::partial$templates/test.hbs[]

or, e.g:

{{#each items}}
* {{name}}
{{/each}}
--
```

Supports

* https://issues.couchbase.com/browse/DOC-10720 (Server Metrics)

* https://github.com/couchbase/docs-sync-gateway/pull/738 (SGW JSON configuration docs)
Using the `each-openapi` helper

* YAML or JSON datafile

* Helpers and Partials 

* Standardised Logging via Dan-approved code 
(though this is via Opal, may be cleaned up in future if exposed natively by Asciidoctor.js)


* Move OpenAPI config gen to JS

e.g. instead of attempting to do it in Handlebars

* Fix each-openapi templating

* markdown-block now uses a library 
(as the markdown in the new OpenAPI spec benefits from this)

NB: for the Docs-site demo, we use an OpenAPI Spec in JSON rather
than YAML.
Though openapi spec defaults to yaml, yaml's pathological
flexibility with references means that, even with
`redocly bundle --defererenced` the definition of schemas
is spread across the file.
Outputting as JSON means that you get a flatter (and much
larger file). The advantage is that it's easier to fillet
(e.g. remove all the bits that aren't in the database schema)
to make it a smaller example file for docs-site.