Minimal reproduction of Antora logging error
It looks like if too many errors are sent to the Antora logger,
it doesn't finish logging all of them.

Not sure if this is because messages aren't drained before Antora
shuts down, or that the logger is otherwise unresponsive.

Case 1: --log-level=fatal
All the messages are seen:

    ❯ antora minimal.yml --extension=lib/minimal.js --log-level=fatal
    [08:34:26.425] FATAL (andora): Context Started (antora logger)
    Context Started (console.log)
    ...
    Site generation complete!
    Open file:///Users/hakimcassimally/couchbase/docs-site/public in a browser to view your site.
    [08:35:10.242] FATAL (andora): Site Published (antora logger)
    Site Published (console.log)

Case 2: --log-level=info
The on('sitePublished') hook is still run (as we see the console.log)
however we DO NOT see the FATAL message via the Antora logger.

    ❯ antora minimal.yml --extension=lib/minimal.js --log-level=info
    [08:36:21.272] FATAL (andora): Context Started (antora logger)
    Context Started (console.log)
    ... (a million errors)
    Site generation complete!
    Open file:///Users/hakimcassimally/couchbase/docs-site/public in a browser to view your site.
    Site Published (console.log)