AV-57335 POC for analytics include using collector
To simplify process for Analytics team, we can use
https://gitlab.com/antora/antora-collector-extension/

NOTE: this extension is currently Alpha, will have to discuss
with OpenDevise.

This PR will add the Couchbase fork of AsterixDB and the Collector
extension.
A subsequent PR on the AsterixDB repo will provide a command to
aggregate sources.

The .adoc file in this PR includes one of these files, using the
Markdown block (modified to cope with the use of HTML in AsterixDB
sources)