DOC-13194, DOC-12510: improve openapi/JSON Schema handling (#823)
* DOC-13194, DOC-12510: improve openapi/JSON Schema handling

* infer type: object if appropriate
Though it would be convenient for us if all nodes were tagged,
redocly doesn't declare that as an issue with the spec, so
we should handle it.

* descend into `allOf`

* move each-openapi.js into helpers

* DOC-13194 DOC-12510: test each-openapi

Run `npm test` from the root of docs-site to test

* DOC-13194 additional tests via @torcolvin

* tidying

* DOC-13194 simple output test for extensions content.

* DOC-13194 simplify

* DOC-13194 npm-test github action

basic github action

* DOC-13194 add playbook for testing

* debug

* fix tests

* fix tests better

I'd not understood the Mocha logic of when code executes...
e.g.
the `ok` blocks are run at the equivalent of "compile time".
They then register the *intent* to run the test *function*.

So we need to do the file read *inside* the registered function.

* indentation