Update Antora preview playbook to use `adding-anywhere-footnote` branch
Add `asciidoctor-anywhere-footnote` to Antora playbook and update UI bundle to prod-213

Add `asciidoctor-anywhere-footnote` to Antora playbook and update UI bundle path for local build

Add `anywhere-footnotes` example and link to navigation

Upgrade `asciidoctor-anywhere-footnote` to ~1.2.7 and update dependencies in `package-lock.json`.

Upgrade `asciidoctor-anywhere-footnote` to ~1.2.2 and update dependencies in `package-lock.json`.

Upgrade `asciidoctor-anywhere-footnote` to ~1.2.1 and update dependencies in `package-lock.json`.

Upgrade `asciidoctor-anywhere-footnote` to ~1.2.0 and update dependencies in `package-lock.json`.

Add `asciidoctor-anywhere-footnote` to Antora preview playbook

Upgrade `asciidoctor-anywhere-footnote` to ~1.1.9 and update dependencies in `package-lock.json`.

Updated to include the correct URL for the UI bundle.

The PR for the new UI bundle (https://github.com/couchbase/docs-ui/pull/193) should be reviewed first so we know the number of the bundle.

Upgrade `asciidoctor-anywhere-footnote` to ~1.1.8 and update dependencies in `package-lock.json`.

Updating package.json

Adding the anywhere-footnote extension