Upgrade `asciidoctor-anywhere-footnote` to ~1.3.5 in `package.json`.
Update footnote markers in `anywhere-footnotes` example to remove brackets and improve readability.

Update UI bundle URL to `adding-anywhere-footnote-13` in Antora preview playbook

Upgrade `asciidoctor-anywhere-footnote` to ~1.3.2 and update dependencies in `package-lock.json`.

Upgrade `asciidoctor-anywhere-footnote` to ~1.3.1, update footnote structure in `anywhere-footnotes` example, and remove the unused `randomstring` dependency

Add `refid` attributes to footnotes in `anywhere-footnotes` example.

This is to ensure that the unit tests don't need to worry about the random strings that are usually applied to the footnote references.