DOC-13521 calculate staging playbook (#840)
* DOC-13521 calculate staging playbook

Rationale: as well as the staging playbook, we now have to manage:

* the template preview playbook
* 1+ playbooks for chatbot
* escrow playbook

All of which need to be kept in sync with the main playbook,
preserving just important differences.

As the staging playbook is the most high-profile one with likely the most
complexity in setting it up, I'm starting with that, to check that we can
do this fully enough to support the requirement.

(The preview one is currently supported by the more straight-forward
scripts/patch-preview-playbook, but could use the same code.)

* add the staging.diff playbook :facepalm: